.circleci: Add slash to end of s3 cp (#43792)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43792

This fixes the issue we had with the nightlies not being uploaded
properly, basically what was happening was that `aws s3 cp` doesn't
automatically distinguish between prefixes that are already
"directories" vs a single file with the same name.

This means that if you'd like to upload a file to a "directory" in S3
you need to suffix your destination with a slash.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D23402074

Pulled By: seemethere

fbshipit-source-id: 6085595283fcbbbab0836ccdfe0f8aa2a6abd7c8